доработал view
