+++ src/presenter/board-presenter.js
@@ -1,23 +1,25 @@
-import EventEditFormView from '../view/new-event-edit-view';
-import EventFormView from '../view/new-event-form-view';
-import NewTravelButtonView from '../view/new-travel-button-view';
-import NewFiltres from '../view/new-filtres-view';
-import NewSorts from '../view/new-sort-view';
-import CreatePointView from '../view/new-point-travel-view';
-import {render} from '../render.js';
+import EventSortView from '../view/event-edit-form-view.js';
+// import EventFormView from '../view/new-event-form-view';
+// import NewTravelButtonView from '../view/new-travel-button-view';
+// import NewFiltres from '../view/new-filtres-view';
+// import SortView from '../view/new-sort-view';
+// import CreatePointView from '../view/new-point-travel-view';
+import {render} from '../render';
 
 export default class BoardPresenter {
-  headerContainer = document.querySelector('.page-header__container');
+  containerElement = null;
 
-  init = (headerContainer) => {
-    this.headerContainer = headerContainer;
+  init = (containerElement) => {
+    this.containerElement = containerElement; // .trip-events (<section class="trip-events">)
 
-    render(new EventEditFormView(), this.headerContainer.getElement());
-    render(new EventFormView(), this.headerContainer.getElement());
-    render(new NewTravelButtonView(), this.headerContainer.getElement());
-    render(new NewFiltres(), this.headerContainer.getElement());
-    render(new NewSorts(), this.headerContainer.getElement());
-    render(new CreatePointView(), this.headerContainer.getElement());
+    render(new EventSortView(), this.containerElement, render);
+
+    // render(new EventEditFormView(), this.headerContainer);
+    // render(new EventFormView(), this.headerContainer.getElement());
+    // render(new NewTravelButtonView(), this.headerContainer.getElement());
+    // render(new NewFiltres(), this.headerContainer.getElement());
+    // render(new SortView(), this.headerContainer);
+    // render(new CreatePointView(), this.headerContainer.getElement());
 
     // for (let i = 0; i < 3; i++) {
     //   render(new TaskView(), this.taskListComponent.getElement());
