+++ src/view/event-edit-form-view.js
@@ -1,8 +1,9 @@
-//event--edit
 import {createElement} from '../render.js';
 
-const EventFormView = () => `<form class="event event--edit" action="#" method="post">
-<header class="event__header">
+const createMainTemplate = () => (
+  `<li class="trip-events__item">
+  <form class="event event--edit" action="#" method="post">
+  <header class="event__header">
   <div class="event__type-wrapper">
     <label class="event__type  event__type-btn" for="event-type-toggle-1">
       <span class="visually-hidden">Choose event type</span>
@@ -66,7 +67,7 @@ const EventFormView = () => `<form class="event event--edit" action="#" method="
     <label class="event__label  event__type-output" for="event-destination-1">
       Flight
     </label>
-    <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Geneva" list="destination-list-1">
+    <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Chamonix" list="destination-list-1">
     <datalist id="destination-list-1">
       <option value="Amsterdam"></option>
       <option value="Geneva"></option>
@@ -76,10 +77,10 @@ const EventFormView = () => `<form class="event event--edit" action="#" method="
 
   <div class="event__field-group  event__field-group--time">
     <label class="visually-hidden" for="event-start-time-1">From</label>
-    <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="19/03/19 00:00">
+    <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="18/03/19 12:25">
     &mdash;
     <label class="visually-hidden" for="event-end-time-1">To</label>
-    <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="19/03/19 00:00">
+    <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="18/03/19 13:35">
   </div>
 
   <div class="event__field-group  event__field-group--price">
@@ -87,11 +88,14 @@ const EventFormView = () => `<form class="event event--edit" action="#" method="
       <span class="visually-hidden">Price</span>
       &euro;
     </label>
-    <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="">
+    <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="160">
   </div>
 
   <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-  <button class="event__reset-btn" type="reset">Cancel</button>
+  <button class="event__reset-btn" type="reset">Delete</button>
+  <button class="event__rollup-btn" type="button">
+    <span class="visually-hidden">Open event</span>
+  </button>
 </header>
 <section class="event__details">
   <section class="event__section  event__section--offers">
@@ -103,16 +107,16 @@ const EventFormView = () => `<form class="event event--edit" action="#" method="
         <label class="event__offer-label" for="event-offer-luggage-1">
           <span class="event__offer-title">Add luggage</span>
           &plus;&euro;&nbsp;
-          <span class="event__offer-price">30</span>
+          <span class="event__offer-price">50</span>
         </label>
       </div>
 
       <div class="event__offer-selector">
         <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked>
         <label class="event__offer-label" for="event-offer-comfort-1">
-          <span class="event__offer-title">Switch to comfort class</span>
+          <span class="event__offer-title">Switch to comfort</span>
           &plus;&euro;&nbsp;
-          <span class="event__offer-price">100</span>
+          <span class="event__offer-price">80</span>
         </label>
       </div>
 
@@ -147,24 +151,17 @@ const EventFormView = () => `<form class="event event--edit" action="#" method="
 
   <section class="event__section  event__section--destination">
     <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-    <p class="event__destination-description">Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.</p>
-
-    <div class="event__photos-container">
-      <div class="event__photos-tape">
-        <img class="event__photo" src="img/photos/1.jpg" alt="Event photo">
-        <img class="event__photo" src="img/photos/2.jpg" alt="Event photo">
-        <img class="event__photo" src="img/photos/3.jpg" alt="Event photo">
-        <img class="event__photo" src="img/photos/4.jpg" alt="Event photo">
-        <img class="event__photo" src="img/photos/5.jpg" alt="Event photo">
-      </div>
-    </div>
+    <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
   </section>
 </section>
-</form>`;
+</form>
+</li>`
+);
 
-export default class CreateEventFormView {
+// new-event-edit-view !== EventEditFormView
+export default class EventEditFormView {
   getTemplate() {
-    return EventFormView();
+    return createMainTemplate();
   }
 
   getElement() {
