+++ src/view/event-filtres-view.js
@@ -1,6 +1,10 @@
 import {createElement} from '../render.js';
 
-const FiltresView = () => `<form class="trip-filters" action="#" method="get">
+const createTemplate = () => (`
+<div class="trip-main__trip-controls  trip-controls">
+<div class="trip-controls__filters">
+  <h2 class="visually-hidden">Filter events</h2>
+  <form class="trip-filters" action="#" method="get">
 <div class="trip-filters__filter">
   <input id="filter-everything" class="trip-filters__filter-input  visually-hidden" type="radio" name="trip-filter" value="everything">
   <label class="trip-filters__filter-label" for="filter-everything">Everything</label>
@@ -17,11 +21,15 @@ const FiltresView = () => `<form class="trip-filters" action="#" method="get">
 </div>
 
 <button class="visually-hidden" type="submit">Accept filter</button>
-</form>`;
+</form>
+</div>
+</div>
+`
+);
 
-export default class CreateFiltresView {
+export default class EventFiltresView {
   getTemplate() {
-    return FiltresView();
+    return createTemplate();
   }
 
   getElement() {
