+++ src/view/event-form-view.js
@@ -1,6 +1,9 @@
+//event--edit
 import {createElement} from '../render.js';
 
-const createNewEventForm = () => `<form class="event event--edit" action="#" method="post">
+const createTemplate = () => (
+  `<li class="trip-events__item">
+  <form class="event event--edit" action="#" method="post">
 <header class="event__header">
   <div class="event__type-wrapper">
     <label class="event__type  event__type-btn" for="event-type-toggle-1">
@@ -65,7 +68,7 @@ const createNewEventForm = () => `<form class="event event--edit" action="#" met
     <label class="event__label  event__type-output" for="event-destination-1">
       Flight
     </label>
-    <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Chamonix" list="destination-list-1">
+    <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="Geneva" list="destination-list-1">
     <datalist id="destination-list-1">
       <option value="Amsterdam"></option>
       <option value="Geneva"></option>
@@ -75,10 +78,10 @@ const createNewEventForm = () => `<form class="event event--edit" action="#" met
 
   <div class="event__field-group  event__field-group--time">
     <label class="visually-hidden" for="event-start-time-1">From</label>
-    <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="18/03/19 12:25">
+    <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="19/03/19 00:00">
     &mdash;
     <label class="visually-hidden" for="event-end-time-1">To</label>
-    <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="18/03/19 13:35">
+    <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="19/03/19 00:00">
   </div>
 
   <div class="event__field-group  event__field-group--price">
@@ -86,14 +89,11 @@ const createNewEventForm = () => `<form class="event event--edit" action="#" met
       <span class="visually-hidden">Price</span>
       &euro;
     </label>
-    <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="160">
+    <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="">
   </div>
 
   <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-  <button class="event__reset-btn" type="reset">Delete</button>
-  <button class="event__rollup-btn" type="button">
-    <span class="visually-hidden">Open event</span>
-  </button>
+  <button class="event__reset-btn" type="reset">Cancel</button>
 </header>
 <section class="event__details">
   <section class="event__section  event__section--offers">
@@ -105,16 +105,16 @@ const createNewEventForm = () => `<form class="event event--edit" action="#" met
         <label class="event__offer-label" for="event-offer-luggage-1">
           <span class="event__offer-title">Add luggage</span>
           &plus;&euro;&nbsp;
-          <span class="event__offer-price">50</span>
+          <span class="event__offer-price">30</span>
         </label>
       </div>
 
       <div class="event__offer-selector">
         <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox" name="event-offer-comfort" checked>
         <label class="event__offer-label" for="event-offer-comfort-1">
-          <span class="event__offer-title">Switch to comfort</span>
+          <span class="event__offer-title">Switch to comfort class</span>
           &plus;&euro;&nbsp;
-          <span class="event__offer-price">80</span>
+          <span class="event__offer-price">100</span>
         </label>
       </div>
 
@@ -149,14 +149,27 @@ const createNewEventForm = () => `<form class="event event--edit" action="#" met
 
   <section class="event__section  event__section--destination">
     <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-    <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps, it's renowned for its skiing.</p>
+    <p class="event__destination-description">Geneva is a city in Switzerland that lies at the southern tip of expansive Lac Léman (Lake Geneva). Surrounded by the Alps and Jura mountains, the city has views of dramatic Mont Blanc.</p>
+
+    <div class="event__photos-container">
+      <div class="event__photos-tape">
+        <img class="event__photo" src="img/photos/1.jpg" alt="Event photo">
+        <img class="event__photo" src="img/photos/2.jpg" alt="Event photo">
+        <img class="event__photo" src="img/photos/3.jpg" alt="Event photo">
+        <img class="event__photo" src="img/photos/4.jpg" alt="Event photo">
+        <img class="event__photo" src="img/photos/5.jpg" alt="Event photo">
+      </div>
+    </div>
   </section>
 </section>
-</form>`;
+</form>
+</li>
+`
+);
 
-export default class EventEditFormView {
+export default class EventFormView {
   getTemplate() {
-    return EventFormView();
+    return createTemplate();
   }
 
   getElement() {
